--- perfume-api/src/docs/asciidoc/index.adoc
@@ -1,3 +1,0 @@
-include::{snippets}/index/http-request.adoc[]
-include::{snippets}/index/http-response.adoc[]
-include::{snippets}/index/curl-request.adoc[]
