+++ perfume-api/src/test/java/io/perfume/api/auth/stub/StubAuthenticationKeyRepository.java
@@ -1,6 +1,6 @@
 package io.perfume.api.auth.stub;
 
-import io.perfume.api.auth.adapter.out.persistence.persistence.AuthenticationKeyRepository;
+import io.perfume.api.auth.application.port.out.AuthenticationKeyRepository;
 import io.perfume.api.auth.domain.AuthenticationKey;
 
 import java.util.ArrayList;
